--- a/src/connectivity.jl
+++ b/src/connectivity.jl
@@ -505,11 +505,13 @@ neighborhood_dists(g::AbstractGraph{T}, v::Integer, d, distmx::AbstractMatrix{U}
 
 
 function _neighborhood(g::AbstractGraph{T}, v::Integer, d::Real, distmx::AbstractMatrix{U}, neighborfn::Function) where T <: Integer where U <: Real
-    d < 0 && return Vector{Tuple{T,U}}()
+    ud = U(d)
+    if ud < 0
+        return Vector{Tuple{T, U}}()
+    end
+
+    #= ud < zero(U) && return Vector{Tuple{T,U}}() =#
     neighs = Vector{T}()
-    neighset = Set{T}()
-    push!(neighs, v)
-    push!(neighset, v)
     Q = Vector{T}()
     push!(Q, v)
     seen = falses(nv(g))
@@ -520,14 +522,13 @@ function _neighborhood(g::AbstractGraph{T}, v::Integer, d::Real, distmx::Abstrac
         seen[src] && continue
         seen[src] = true
         currdist = dists[src]
+        push!(neighs, src)
         vertexneighbors = neighborfn(g, src)
         @inbounds for vertex in vertexneighbors
             if !seen[vertex] 
                 vdist = currdist == typemax(U) ? typemax(U) : currdist + distmx[src, vertex]
-                if vdist <= d && vertex ∉ neighset
+                if vdist <= ud
                     push!(Q, vertex)
-                    push!(neighs, vertex)
-                    push!(neighset, vertex)
                     dists[vertex] = vdist
                 end
             end