This has a bad effect on (worst-case) performance:

```
julia> g = CompleteGraph(5000)
{5000, 12497500} undirected simple Int64 graph
```
OLD:
```
julia> @benchmark neighborhood_dists(g, 1, 4)
BenchmarkTools.Trial: 
  memory estimate:  1.96 MiB
  allocs estimate:  40440
  --------------
  minimum time:     579.565 ms (0.00% GC)
  median time:      589.257 ms (0.00% GC)
  mean time:        592.398 ms (0.00% GC)
  maximum time:     617.377 ms (0.00% GC)
  --------------
  samples:          9
  evals/sample:     1
```

NEW:
```
julia> @benchmark neighborhood_dists(g, 1, 4)
BenchmarkTools.Trial: 
  memory estimate:  779.59 KiB
  allocs estimate:  13524
  --------------
  minimum time:     1.142 s (0.00% GC)
  median time:      1.155 s (0.00% GC)
  mean time:        1.162 s (0.00% GC)
  maximum time:     1.189 s (0.00% GC)
  --------------
  samples:          5
  evals/sample:     1
```