@sbromberger 2019-03-08
Making `neighset` a bitvector:
```
julia> g = CompleteGraph(5000)
{5000, 12497500} undirected simple Int64 graph

julia> @benchmark neighborhood_dists(g, 1, 4)
BenchmarkTools.Trial: 
  memory estimate:  586.48 KiB
  allocs estimate:  13504
  --------------
  minimum time:     1.014 s (0.00% GC)
  median time:      1.020 s (0.00% GC)
  mean time:        1.021 s (0.00% GC)
  maximum time:     1.031 s (0.00% GC)
  --------------
  samples:          5
  evals/sample:     1
```