@sbromberger 2019-03-08
With `Vector{Bool}`:
```
julia> @benchmark neighborhood_dists(g, 1, 4)
BenchmarkTools.Trial: 
  memory estimate:  600.67 KiB
  allocs estimate:  13504
  --------------
  minimum time:     1.075 s (0.00% GC)
  median time:      1.079 s (0.00% GC)
  mean time:        1.081 s (0.00% GC)
  maximum time:     1.089 s (0.00% GC)
  --------------
  samples:          5
  evals/sample:     1
```