@sbromberger 2019-03-09
now with type stability and no extraneous cruft:
```
julia> @benchmark neighborhood_dists(g, 1, 4)
BenchmarkTools.Trial: 
  memory estimate:  375.30 KiB
  allocs estimate:  34
  --------------
  minimum time:     100.001 ms (0.00% GC)
  median time:      146.391 ms (0.00% GC)
  mean time:        144.894 ms (0.00% GC)
  maximum time:     159.513 ms (0.00% GC)
  --------------
  samples:          35
  evals/sample:     1
```